+++ icekit/pages_api/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = '%s.apps.PagesAPIConfig' % __name__