+++ icekit/project/urls.py
@@ -40,8 +40,8 @@ urlpatterns = patterns(
         {'sitemaps': sitemaps}),
 
     # Installed apps.
-    url(r'^api/pages/', include('icekit.pages_api.urls')),
     url(r'^forms/', include('forms_builder.forms.urls')),
+    url(r'^api/', include('icekit.api.urls')),
 
     # Get admin URLs prefix from settings.
     # Handle admin and front-end authentication separately.