+++ icekit/tests/tests.py
@@ -539,39 +539,6 @@ class Views(WebTest):
         response = self.client.get(self.layoutpage_1.get_absolute_url())
         self.assertEqual(response.status_code, 200)
 
-    def test_page_api(self):
-        self.layoutpage_1.publish()
-        for j in range(20):
-            published_layoutpage = LayoutPage.objects.create(
-                author=self.super_user_1,
-                title='Test LayoutPage %s' % j,
-                layout=self.layout_1,
-            )
-            published_layoutpage.publish()
-
-            draft_layoutpage = LayoutPage.objects.create(
-                author=self.super_user_1,
-                title='Draft LayoutPage %s' % j,
-                layout=self.layout_1,
-            )
-
-        response = self.client.get(reverse('page-list'))
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(len(response.data['results']), 5)
-        self.assertEqual(response.data['count'], 21)
-        response = self.client.get(reverse('page-detail', args=(self.layoutpage_1.id,)))
-        self.assertEqual(response.status_code, 404)
-        response = self.client.get(reverse('page-detail', args=(self.layoutpage_1.get_published().id,)))
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(len(response.data), 6)
-        for key in response.data.keys():
-            if key is not 'content':
-                self.assertEqual(response.data[key],
-                                 getattr(self.layoutpage_1.get_published(), key))
-        self.assertEqual(len(response.data['content']), 3)
-        for number, item in enumerate(response.data['content'], 1):
-            self.assertEqual(item['content'], getattr(self, 'content_instance_%s' % number).html)
-
 
 class TestValidators(WebTest):
     def test_template_name(self):